Firm show ready

+++ app/controllers/firms_controller.rb
@@ -17,6 +17,16 @@ class FirmsController < ApplicationController
     render({ :template => "firms/show.html.erb" })
   end
 
+  def edit
+    the_id = params.fetch("path_id")
+
+    matching_firms = Firm.where({ :id => the_id })
+
+    @the_firm = matching_firms.at(0)
+
+    render({ :template => "firms/edit.html.erb" })
+  end
+
   def create
     the_firm = Firm.new
     the_firm.firm_name = params.fetch("query_firm_name")
@@ -24,7 +34,7 @@ class FirmsController < ApplicationController
 
     if the_firm.valid?
       the_firm.save
-      redirect_to("/firms", { :notice => "Firm created successfully." })
+      redirect_to("/firms", { :notice => "Company created successfully." })
     else
       redirect_to("/firms", { :alert => the_firm.errors.full_messages.to_sentence })
     end
@@ -35,11 +45,22 @@ class FirmsController < ApplicationController
     the_firm = Firm.where({ :id => the_id }).at(0)
 
     the_firm.firm_name = params.fetch("query_firm_name")
-    the_firm.user_id = params.fetch("query_user_id")
+    the_firm.user_id = @current_user.id
 
     if the_firm.valid?
       the_firm.save
-      redirect_to("/firms/#{the_firm.id}", { :notice => "Firm updated successfully."} )
+      
+      the_firm.contacts.each do |a_contact|
+        a_contact.firm_name = the_firm.firm_name
+        a_contact.save
+      end
+
+      the_firm.jobs.each do |a_job|
+        a_job.firm_name = the_firm.firm_name
+        a_job.save
+      end
+
+      redirect_to("/firms/#{the_firm.id}", { :notice => "Company updated successfully."} )
     else
       redirect_to("/firms/#{the_firm.id}", { :alert => the_firm.errors.full_messages.to_sentence })
     end
@@ -49,8 +70,13 @@ class FirmsController < ApplicationController
     the_id = params.fetch("path_id")
     the_firm = Firm.where({ :id => the_id }).at(0)
 
-    the_firm.destroy
-
-    redirect_to("/firms", { :notice => "Firm deleted successfully."} )
+    if the_firm.contacts.count == 0 && the_firm.jobs.count == 0
+      the_firm.destroy
+      redirect_to("/firms", { :notice => "Company deleted successfully."} )
+    else
+      redirect_to("/firms/#{the_firm.id}", { :alert => "You can only delete a company if it has 0 Job Applications and 0 Contacts."} )
+    end
   end
+
+
 end
